fix: make update-manifest pre-release, do not delete assets which not dup

+++ src/lib/release.ts
@@ -6,7 +6,7 @@ import { default as glob } from "fast-glob";
 import fs from "fs-extra";
 import mime from "mime";
 import { Octokit } from "octokit";
-import path from "path";
+import { basename, join } from "path";
 import _ from "radash";
 import { isCI } from "std-env";
 
@@ -57,7 +57,9 @@ export default class Release extends Base {
    * release: bump version, run build, git add, git commit, git tag, git push
    */
   async bump() {
-    await versionBump(this.ctx.release.bumpp);
+    const result = await versionBump(this.ctx.release.bumpp);
+    this.ctx.version = result.newVersion;
+    this.ctx.release.bumpp.tag = result.tag;
     // const releaseItConfig: ReleaseItConfig = {
     //   "only-version": true,
     // };
@@ -96,10 +98,7 @@ export default class Release extends Base {
 
     this.logger.debug("Uploading xpi asset...");
 
-    await this.uploadAsset(
-      release.id,
-      path.join(this.dist, `${this.xpiName}.xpi`),
-    );
+    await this.uploadAsset(release.id, join(this.dist, `${this.xpiName}.xpi`));
   }
 
   async getReleaseByTag(tag: string) {
@@ -148,7 +147,7 @@ export default class Release extends Base {
           "content-type": mime.getType(asset) || "application/octet-stream",
           "content-length": fs.statSync(asset).size,
         },
-        name: path.basename(asset),
+        name: basename(asset),
       })
       .then((res) => {
         return res.data;
@@ -156,7 +155,7 @@ export default class Release extends Base {
   }
 
   async refreshUpdateManifest() {
-    const assets = glob.globSync(`${this.dist}/*.json`);
+    const assets = glob.globSync(`${this.dist}/*.json`).map((p) => basename(p));
 
     const release =
       (await this.getReleaseByTag("release")) ??
@@ -164,19 +163,11 @@ export default class Release extends Base {
         owner: this.owner,
         repo: this.repo,
         tag_name: "release",
+        prerelease: true,
         make_latest: "false",
       }));
 
     if (!release) throw new Error("Get or create 'release' failed.");
-
-    await this.client.rest.repos.updateRelease({
-      owner: this.owner,
-      repo: this.repo,
-      release_id: release.id,
-      name: "Release Manifest",
-      body: `This release is used to host \`update.json\`, please do not delete or modify it! \n Updated in UTC ${new Date().toISOString()} for version ${this.version}`,
-      make_latest: "false",
-    });
 
     const existAssets = await this.client.rest.repos
       .listReleaseAssets({
@@ -190,17 +181,29 @@ export default class Release extends Base {
 
     if (existAssets) {
       for (const existAsset of existAssets) {
-        await this.client.rest.repos.deleteReleaseAsset({
-          owner: this.owner,
-          repo: this.repo,
-          asset_id: existAsset.id,
-        });
+        if (assets.includes(existAsset.name)) {
+          await this.client.rest.repos.deleteReleaseAsset({
+            owner: this.owner,
+            repo: this.repo,
+            asset_id: existAsset.id,
+          });
+        }
       }
     }
 
     for (const asset of assets) {
-      await this.uploadAsset(release.id, path.join(this.dist, asset));
+      await this.uploadAsset(release.id, join(this.dist, asset));
     }
+
+    await this.client.rest.repos.updateRelease({
+      owner: this.owner,
+      repo: this.repo,
+      release_id: release.id,
+      name: "Release Manifest",
+      body: `This release is used to host \`update.json\`, please do not delete or modify it! \n Updated in UTC ${new Date().toISOString()} for version ${this.version}`,
+      prerelease: true,
+      make_latest: "false",
+    });
   }
 
   getChangelog(): Promise<string> {
